Fixing Qt detection of CPU architecture, adding armv6l detection
Previously, compilation worked for all architectures at the
expense of the use of SSE2 instructions. The offending line
(where $$QMAKE_HOST.arch is checked) always returned true. This
is because $$QMAKE_HOST.arch was not read properly before the
equals() check, making the check return false. It was then not-ed
and returned true. Thus support for ARM was achieved because of
two errors cancelling eachother out, but negatively impacting
non-ARM processor performance.

This patch explicitly defines QMAKE_XCPUARCH = $$QMAKE_HOST.arch
before the equals check, removes the not on the equals() check,
and adds in a check for armv6l architecture CPUs as well.